Refactored some containers and added preview flatlist component
