+++ src/containers/index.ts
@@ -2,4 +2,7 @@ export { default as HomeContainer } from './home/home.container';
 export { default as SplashContainer } from './splash/splash.container';
 export { default as UserContainer } from './user/user.container';
 export { default as PreviewContainer } from './preview/preview.container';
-export { default as FormInputContainer } from './form-input/form-input.container';
+export { default as TextPreviewContainer } from './text-preview/text-preview.container';
+export { default as InputPreviewContainer } from './input-preview/input-preview.container';
+export { default as ButtonPreviewContainer } from './button-preview/button-preview.container';
+export { default as FlatListPreviewContainer } from './flatlist-preview/flatlist-preview.container';
