+++ src/navigators/root.navigator.tsx
@@ -1,19 +1,17 @@
 import React from 'react';
 import { StatusBar } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { createDrawerNavigator } from '@react-navigation/drawer';
 import { NavigationContainer } from '@react-navigation/native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 
 import { useTheme } from '@/hooks';
+import { SplashContainer } from '@/containers';
 
 import { navigationRef } from './utils';
 import { RootStackParamList } from './types';
-import { FormInputContainer, SplashContainer } from '@/containers';
-import AppNavigator from './app.navigator';
+import DrawerNavigator from './drawer.navigator';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
-const Drawer = createDrawerNavigator<RootStackParamList>();
 
 // @refresh reset
 const RootNavigator = () => {
@@ -41,23 +39,6 @@ const RootNavigator = () => {
         </Stack.Navigator>
       </NavigationContainer>
     </GestureHandlerRootView>
-  );
-};
-
-const DrawerNavigator = () => {
-  return (
-    <Drawer.Navigator>
-      <Drawer.Screen
-        name="MainTab"
-        component={AppNavigator}
-        options={{ title: 'RN Boilerplate', headerTitle: 'RN Boilerplate' }}
-      />
-      <Drawer.Screen
-        name="FormInput"
-        component={FormInputContainer}
-        options={{ title: 'Input', headerTitle: 'Input Components' }}
-      />
-    </Drawer.Navigator>
   );
 };
